Update to Version 1.3

+++ Generator/Hints/HintTypes/BeyondPointHint.cs
@@ -36,7 +36,7 @@ namespace TPRandomizer.Hints
             List<string> goodChecks,
             List<string> bigKeyChecks,
             bool bigKeyUseDefiniteArticle = false,
-            Dictionary<int, byte> itemPlacements = null
+            Dictionary<int, int> itemPlacements = null
         )
         {
             this.includeBigKeyInfo = includeBigKeyInfo;
@@ -52,7 +52,7 @@ namespace TPRandomizer.Hints
             CalcDerived(genData, itemPlacements);
         }
 
-        private void CalcDerived(HintGenData genData, Dictionary<int, byte> itemPlacements)
+        private void CalcDerived(HintGenData genData, Dictionary<int, int> itemPlacements)
         {
             buildCheckToItemMappings(goodChecks, itemPlacements);
             buildCheckToItemMappings(bigKeyChecks, itemPlacements);
@@ -81,7 +81,7 @@ namespace TPRandomizer.Hints
 
         private void buildCheckToItemMappings(
             List<string> checkNames,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             if (!ListUtils.isEmpty(checkNames))
@@ -183,7 +183,7 @@ namespace TPRandomizer.Hints
         public static BeyondPointHint decode(
             HintEncodingBitLengths bitLengths,
             BitsProcessor processor,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             bool includeBigKeyInfo = processor.NextBool();
